+++ tests/unit/framework/test_pipeline.py
@@ -58,7 +58,6 @@ class TestPipeline(ConfigurationForTests):
             config=self.logprep_config,
             log_queue=mock.MagicMock(),
             lock=self.lock,
-            used_server_ports=mock.MagicMock(),
         )
 
     def test_pipeline_property_returns_pipeline(self, mock_create):
@@ -465,40 +464,6 @@ class TestPipeline(ConfigurationForTests):
             "This is non critical", {"some": "event"}, None
         )
 
-    def test_http_input_registers_to_shard_dict(self, _):
-        self.pipeline._setup()
-        self.pipeline._input.server.config.port = 9000
-        self.pipeline._used_server_ports = {}
-        self.pipeline._setup()
-        assert 9000 in self.pipeline._used_server_ports
-
-    def test_http_input_registers_increased_port_to_shard_dict(self, _):
-        self.pipeline._setup()
-        self.pipeline._input.server.config.port = 9000
-        self.pipeline._used_server_ports = {9000: "other_process_name"}
-        self.pipeline._setup()
-        assert 9001 in self.pipeline._used_server_ports
-
-    def test_http_input_removes_port_from_shard_dict_on_shut_down(self, _):
-        self.pipeline._setup()
-        self.pipeline._input.server.config.port = 9000
-        self.pipeline._used_server_ports = {}
-        self.pipeline._setup()
-        assert 9000 in self.pipeline._used_server_ports
-        self.pipeline._shut_down()
-        assert 9000 not in self.pipeline._used_server_ports
-
-    def test_http_input_registers_increased_port_to_shard_dict_after_shut_down(self, _):
-        self.pipeline._setup()
-        self.pipeline._input.server.config.port = 9000
-        self.pipeline._used_server_ports = {9000: "other_process_name"}
-        self.pipeline._setup()
-        assert 9001 in self.pipeline._used_server_ports
-        self.pipeline._shut_down()
-        assert 9001 not in self.pipeline._used_server_ports
-        self.pipeline._setup()
-        assert 9001 in self.pipeline._used_server_ports
-
     def test_shut_down_drains_input_queues(self, _):
         self.pipeline._setup()
         input_config = {
@@ -513,7 +478,9 @@ class TestPipeline(ConfigurationForTests):
                 "endpoints": {"/json": "json", "/jsonl": "jsonl", "/plaintext": "plaintext"},
             }
         }
-        self.pipeline._input = original_create(input_config, mock.MagicMock())
+        self.pipeline._input = original_create(input_config, self.pipeline.logger)
+        self.pipeline._input.pipeline_index = 1
+        self.pipeline._input.setup()
         self.pipeline._input.messages.put({"message": "test message"})
         assert self.pipeline._input.messages.qsize() == 1
         self.pipeline._shut_down()